Reactor::wait_for takes `&Pollable` instead of `Pollable`
We want to be able to reuse a pollable on many uses in the reactor.

The fundamental underlying operation `wasi:io/poll.poll` takes a `list<borrow<pollable>>`, or `&[&Pollable]` in Rust.

The `Poller` structure needs to be changed to unsafe transmute the `&Pollable` to a common lifetime for passing to `poll`. This remains safe as long as each caller that inserts a `&Pollable` unconditionally removes it.