Switch to testing as much as possible with `cargo test --platform wasm32-wasip2`
which cargo is configured, in this project, to delegate to a `wasmtime` executable.

The tcp_echo_server example has a nontrivial test harness so it is still tested
using the test-programs/artifacts approach. The test harness was moved into
test-programs-artifacts so that all of the wasmtime dev-dependencies would be
in that crate.

new `#[wstd::test]` macro makes a `#[test]` run in the runtime.

CI installs wasmtime and invokes both sets of tests, wstd's integration tests
and then the test-programs-artifact test.