Switch to the http crate's Request, Response, and StatusCode.
Since as of #19, wstd is using the http crate's types for `HeaderMap`,
`Method`, and `Uri`, switch it to use the http crate's types for `Request`,
`Response`, and `StatusCode` too. These types have a few more features,
and this change makes wstd more interoperable with code that uses the http
crate.

This also helps with #34, as the new `Request` and `Response` types here
are independent of being incoming or outgoing.