Optimize copy.
Optimize `wstd::io::copy` using the wasi-io `splice` function. To do
this, add a way to query `AsyncRead` and `AsyncWrite` implementations
for an underlying `AsyncInputStream` and `AsyncOutputStream`.

This is similar to optimizations done in `std::io::copy`, however
those optimizations make use of specialization, which isn't available
on stable Rust. That said, using explicit interfaces instead of
specialization does have the advantage that user-defined types can
opt into optimization if they're able to.

This also eliminates the `RefCell`s in `TcpStream`.