Reinstate `#![forbid(unsafe_code)]`.
The README.md advertises that this crate uses `deny(unsafe_code)`;
change this to say `forbid(unsafe_code)` assuming this is what was
intended, and reinstate the `#![forbid(unsafe_code)]`.

To do this, eliminate an `unsafe` block for creating a noop waker.
This requires doing an extra heap allocation for now, though that
will go away when `task::Waker::noop` is stabilized.