Allow futures independent of wasi pollables to progress
This commit moves away from using noop waker for the main task.
Instead, it keeps track if the main task is ready to progress again
immediately after `Future::poll()`.

If it is, the runtime will skip calling `wasi::io::poll::poll` if there
are no wasi pollables scheduled, or otherwise call in a non-blocking
manner to quickly get an update for all wasi pollables before
progressing the main task again.