Add a new test that uses futures that do not pend on wasi pollables
This test demonstrates behavior for a future which does not pend on a
pollable, but does wake the root task when pending. As long as the
root task has had wake() called, the runtime should keep polling it.

This bug was reported in https://github.com/yoshuawuyts/wstd/issues/69
and this test was derived from one written by @SilverMira in
https://github.com/SilverMira/wstd/tree/allow_independent_futures

Co-Authored-By: SilverMira <66930495+SilverMira@users.noreply.github.com>