Allow non-pollable woken futures to make progress (#71)
* Add a new test that uses futures that do not pend on wasi pollables

This test demonstrates behavior for a future which does not pend on a
pollable, but does wake the root task when pending. As long as the
root task has had wake() called, the runtime should keep polling it.

This bug was reported in https://github.com/yoshuawuyts/wstd/issues/69
and this test was derived from one written by @SilverMira in
https://github.com/SilverMira/wstd/tree/allow_independent_futures

Co-Authored-By: SilverMira <66930495+SilverMira@users.noreply.github.com>

* track root task's awake status, and skip blocking on pollables if awake

* CI: run cargo test with --nocapture so i can see the assertion that failed

* fix windows CI

* comment!

---------

Co-authored-by: SilverMira <66930495+SilverMira@users.noreply.github.com>